refactor(ffi): restructure the package around the isolate boundary
The layering existed in the doc comments but not in the tree: io/ mixed
engine-isolate and root-isolate files, engine_isolate.dart held both sides
of the port (EngineHost on the root isolate, the frame driver on the engine
one), and view/ imported the engine when it only needed the protocol.

lib/src is now three layers plus the two doors:

  protocol/       sendable message types only, including the query envelopes
  engine/         EngineHost, MapSession, render backend; engine/core/ is the
                  isolate itself and the only owner of mln.*
  presentation/   platform adapter, map widget, gestures, ornaments
  native_bridge.dart  the single door to Kotlin (was io/texture_bridge.dart,
                      renamed because it also does location, cache and links)
  utils/projection.dart  Mercator and camera limits, mirroring mbgl::util

The two god files are split by concern. ffi_platform.dart goes from 1706 to
1093 lines, keeping all 96 platform-interface overrides as delegation:

  style_layers.dart      a StyleLayerType enum plus one JSON builder, so the
                         8 addXLayer methods share a body and no longer
                         repeat the style-spec type as a literal
  camera_update_codec.dart  CameraUpdate -> EngineCommand, pure
  map_options.dart       every maplibre_gl option key in one place
  feature_interaction.dart  hit-testing, tap and drag payloads
  location_component.dart   the whole myLocation subsystem
  snapshot_service.dart  request-id registry and PNG encoding
  image_codec.dart       image decode and fetch

ornaments.dart becomes one file per ornament, with an OrnamentPosition enum
replacing the magic 0..3 indexes and an OrnamentsOverlay that absorbs the
logo/attribution collision rule; ffi_map_view.dart drops from 507 to 405
lines, with the recognizer wiring in map_gesture_detector.dart.

MapSession replaces the (host, sessionId) threading: 41 tuple destructurings
and the gesture handler's two injected closures collapse into one handle.

Fixes found on the way:
- getMetersPerPixelAtLatitude had two divergent copies of the formula, one
  with an imprecise earth circumference, and neither applied the latitude and
  zoom clamps mbgl::Projection applies; it now matches Android and iOS
  outside the Mercator range. First tests in the package cover it.
- takeSnapshot had lost its explicit width/height on the platform side.
- FfiMapView.dispose() unsubscribed from engine events after disposing the
  notifiers its handler writes to.
- The 32 "// --- Section ---" banners, 5 of which marked nothing after an
  earlier part-file split, become plain one-line comments. Convention and
  the new layout are documented in ARCHITECTURE.md.

No behavior change beyond the projection clamps and the takeSnapshot fix.
