+++ maplibre_gl_native/lib/src/engine/core/engine_core_commands.dart
