+++ maplibre_gl_native/lib/src/engine/core/frame_stats.dart
