+++ maplibre_gl_native/lib/src/engine/core/http_resource_provider.dart
