+++ maplibre_gl_native/lib/src/engine/core/vsync_pulse.dart
