+++ maplibre_gl_native/lib/src/presentation/platform/ffi_platform_base.dart
