+++ maplibre_gl_native/lib/src/presentation/platform/style_string_resolver.dart
